document_qa_gemini/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
python-multipart
PyPDF2
python-docx
aiofiles
pydantic
mcp
pyee